task-scheduler/pyproject.toml
[project]
name = "task-scheduler"
version = "0.1.0"
description = "Python UV + Prefect Claude Code Task Scheduler"
readme = "README.md"
requires-python = ">=3.9"
dependencies = [
    "prefect>=2.14.0",
    "pyyaml>=6.0",
    "typer>=0.9.0",
    "rich>=13.0.0",
    "python-dateutil>=2.8.0"
]

[dependency-groups]
dev = [
    "pytest>=7.0.0",
    "black>=23.0.0",
    "isort>=5.12.0"
]

[build-system]
requires = ["setuptools>=61.0", "wheel"]
build-backend = "setuptools.build_meta"

[tool.setuptools]
packages = []
py-modules = ["main"]

[tool.uv]
dev-dependencies = [
    "pytest>=7.0.0",
    "black>=23.0.0",
    "isort>=5.12.0"
]

[project.scripts]
task-scheduler = "main:main"
